requirements.txt
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
bcrypt==4.2.0
